# backend/tests/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kivai Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/video/compress/inspect": {
      "post": {
        "summary": "Inspect Video For Compression",
        "operationId": "inspect_video_for_compression_video_compress_inspect_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_inspect_video_for_compression_video_compress_inspect_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/video/compress": {
      "post": {
        "summary": "Compress Video",
        "operationId": "compress_video_video_compress_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_compress_video_video_compress_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/video/hevc-to-mp4": {
      "post": {
        "summary": "Hevc To Mp4",
        "operationId": "hevc_to_mp4_video_hevc_to_mp4_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_hevc_to_mp4_video_hevc_to_mp4_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/video/mp4-to-hevc": {
      "post": {
        "summary": "Mp4 To Hevc",
        "operationId": "mp4_to_hevc_video_mp4_to_hevc_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_mp4_to_hevc_video_mp4_to_hevc_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_compress_video_video_compress_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "balanced"
          },
          "preset": {
            "type": "string",
            "title": "Preset",
            "default": "custom"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "default": "auto"
          },
          "fps": {
            "type": "string",
            "title": "Fps",
            "default": "original"
          },
          "bitrate": {
            "type": "string",
            "title": "Bitrate",
            "default": "auto"
          },
          "custom_bitrate": {
            "type": "integer",
            "title": "Custom Bitrate",
            "default": 2500
          },
          "codec": {
            "type": "string",
            "title": "Codec",
            "default": "h264"
          },
          "audio": {
            "type": "string",
            "title": "Audio",
            "default": "keep"
          },
          "audio_bitrate": {
            "type": "integer",
            "title": "Audio Bitrate",
            "default": 128
          },
          "target_mb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Mb"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_compress_video_video_compress_post"
      },
      "Body_hevc_to_mp4_video_hevc_to_mp4_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_hevc_to_mp4_video_hevc_to_mp4_post"
      },
...
      "Body_inspect_video_for_compression_video_compress_inspect_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_inspect_video_for_compression_video_compress_inspect_post"
      },
...
      "Body_mp4_to_hevc_video_mp4_to_hevc_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_mp4_to_hevc_video_mp4_to_hevc_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}